Renamer pakken; tilstand -> formidling.
Pakken inneholder informasjon om formidling av registrering mot Arena, og ikke tilstand til selve registreringen.
