+++ src/main/java/no/nav/fo/veilarbregistrering/registrering/bruker/BrukerTilstandService.kt
@@ -4,13 +4,12 @@ import no.nav.common.featuretoggle.UnleashService
 import no.nav.fo.veilarbregistrering.bruker.Bruker
 import no.nav.fo.veilarbregistrering.oppfolging.OppfolgingGateway
 import no.nav.fo.veilarbregistrering.registrering.bruker.Resending.kanResendes
-import no.nav.fo.veilarbregistrering.registrering.tilstand.Status
+import no.nav.fo.veilarbregistrering.registrering.formidling.Status
 import no.nav.fo.veilarbregistrering.sykemelding.Maksdato
 import no.nav.fo.veilarbregistrering.sykemelding.SykemeldingService
 import no.nav.fo.veilarbregistrering.sykemelding.SykmeldtInfoData
 import org.slf4j.LoggerFactory
 import kotlin.jvm.JvmOverloads
-import java.time.LocalDateTime
 
 class BrukerTilstandService(
     private val oppfolgingGateway: OppfolgingGateway,
