+++ src/main/java/no/nav/fo/veilarbregistrering/registrering/formidling/OppdaterRegistreringTilstandCommand.java
@@ -1,4 +1,4 @@
-package no.nav.fo.veilarbregistrering.registrering.tilstand;
+package no.nav.fo.veilarbregistrering.registrering.formidling;
 
 public class OppdaterRegistreringTilstandCommand {
