+++ src/main/java/no/nav/fo/veilarbregistrering/registrering/formidling/RegistreringTilstandService.java
@@ -1,4 +1,4 @@
-package no.nav.fo.veilarbregistrering.registrering.tilstand;
+package no.nav.fo.veilarbregistrering.registrering.formidling;
 
 import java.util.List;
