+++ src/main/java/no/nav/fo/veilarbregistrering/registrering/formidling/resources/InternalRegistreringStatusoversiktServlet.java
@@ -1,9 +1,9 @@
-package no.nav.fo.veilarbregistrering.registrering.tilstand.resources;
+package no.nav.fo.veilarbregistrering.registrering.formidling.resources;
 
 import com.google.gson.Gson;
-import no.nav.fo.veilarbregistrering.registrering.tilstand.RegistreringTilstand;
-import no.nav.fo.veilarbregistrering.registrering.tilstand.RegistreringTilstandService;
-import no.nav.fo.veilarbregistrering.registrering.tilstand.Status;
+import no.nav.fo.veilarbregistrering.registrering.formidling.RegistreringTilstand;
+import no.nav.fo.veilarbregistrering.registrering.formidling.RegistreringTilstandService;
+import no.nav.fo.veilarbregistrering.registrering.formidling.Status;
 
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
