+++ src/test/java/no/nav/fo/veilarbregistrering/registrering/formidling/RegistreringTilstandTestdataBuilder.java
@@ -1,4 +1,4 @@
-package no.nav.fo.veilarbregistrering.registrering.tilstand;
+package no.nav.fo.veilarbregistrering.registrering.formidling;
 
 import java.time.LocalDateTime;
